finally get a rendered frame back

+++ src/Core/Render/RIDevice.hpp
@@ -15,7 +15,7 @@ namespace Phoenix
 		virtual ~IRIDevice() {}
 
 		virtual VertexBufferHandle	 createVertexBuffer(const VertexBufferFormat& format) = 0;
-		virtual IndexBufferHandle	 createIndexBuffer(size_t elementSizeBytes, uint32_t count, const void* data) = 0;
+		virtual IndexBufferHandle	 createIndexBuffer(size_t elementSizeBytes, size_t count, const void* data) = 0;
 		virtual VertexShaderHandle	 createVertexShader(const char* source) = 0;
 		virtual FragmentShaderHandle createFragmentShader(const char* source) = 0;
 		virtual ProgramHandle		 createProgram(VertexShaderHandle vs, FragmentShaderHandle fs) = 0;
